+++ src/main/java/com/hotelmangementapi/demo/model/dtos/requests/RoomRequestAndResponse.java
@@ -15,7 +15,7 @@ public class RoomRequestAndResponse {
     private RoomType roomType;
     private  Integer floorNum;
     private  String description;
-    private Integer originalPrice;
+    private Double originalPrice;
 
 
 }
