+++ src/main/java/com/hotelmangementapi/demo/model/dtos/responses/ReservationResponse.java
@@ -17,7 +17,7 @@ public class ReservationResponse {
     private String visitorLastName;
     private LocalDate startingDate;
     private LocalDate endingDate;
-    private Integer reservationPrice;
+    private Double reservationPrice;
 
 
 }
